requirements.txt
alabaster==0.7.16
altgraph==0.17.4
annotated-types==0.7.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
# auto-py-to-exe==2.45.1
babel==2.17.0
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
Brotli==1.1.0
cachelib==0.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
cssselect2==0.8.0
defusedxml==0.7.1
Deprecated==1.2.18
docutils==0.20.1
docx2pdf==0.1.8
Eel==0.18.1
et_xmlfile==2.0.0
Faker==36.1.0
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
Flask-Caching==2.3.0
flask-cors==5.0.1
Flask-Limiter==3.12
fonttools==4.56.0
fsspec==2025.2.0
future==1.0.0
fuzzywuzzy==0.18.0
gevent==24.11.1
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-genai==1.1.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.29.1
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
Levenshtein==0.26.1
limits==5.1.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
mysql-connector-python==9.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
oauthlib==3.2.2
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
ordered-set==4.1.0
oscrypto==1.3.0
packaging==24.2
# pdf2docx==0.5.8
pdfplumber==0.11.6
pdfservices-sdk==4.1.0
pefile==2023.2.7
pillow==11.1.0
playwright==1.50.0
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
pyee==12.1.1
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.9.0
PyMuPDF==1.25.3
PyMySQL==1.1.1
pyparsing==3.2.1
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyphen==0.17.2
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-bidi==0.6.6
python-docx==1.1.2
python-dotenv==1.1.0
python-Levenshtein==0.26.1
python-slugify==8.0.4
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.1
RapidFuzz==3.12.1
redis==5.3.0
regex==2024.11.6
reportlab==4.3.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==69.5.1
six==1.17.0
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
svglib==1.5.1
sympy==1.13.1
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tkcalendar==1.6.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
Unidecode==1.3.8
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
weasyprint==65.0
webencodings==0.5.1
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
xhtml2pdf==0.2.17
XlsxWriter==3.2.2
zope.event==5.0
zope.interface==7.2
zopfli==0.2.3.post1
